Naprawiam po was
nie chce mi sie pisać co zrobiłem bo za dużo
