+++ Front/views/__init__.py
@@ -1,0 +1,1 @@
+from services.imagesService import choose_and_send_photo
